Move row extends into new namespace.
